pyproject.toml
[tool.poetry]
name = "tdeni-github-io"
version = "0.1.0"
description = ""
authors = ["Deni <sha256deni@gmail.com>"]
readme = "README.md"
packages = [{include = "tdeni"}]

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.11.0"}
jinja2 = "^3.1.3"
python-slugify = "^8.0.4"
markdown = "^3.6"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
